Split point coordinates.
