+++ core/specfem/point.hpp
@@ -4,14 +4,15 @@
 #include "point/assembly_index.hpp"
 #include "point/boundary.hpp"
 #include "point/conforming_interface.hpp"
-#include "point/coordinates.hpp"
 #include "point/displacement.hpp"
 #include "point/edge_index.hpp"
 #include "point/field_derivatives.hpp"
+#include "point/global_coordinates.hpp"
 #include "point/index.hpp"
 #include "point/interface_index.hpp"
 #include "point/jacobian_matrix.hpp"
 #include "point/kernels.hpp"
+#include "point/local_coordinates.hpp"
 #include "point/mapped_index.hpp"
 #include "point/mass_inverse.hpp"
 #include "point/nonconforming_interface.hpp"
