+++ core/specfem/point/global_coordinates.cpp
